Split out housekeeping tasks into new emscripten_yield function. NFC
We currently only do two thing in this function:

1. Check if we should die because some other thread crashed.
2. Process the work queue.

However, I have plans to do other things such as keep loaded code
in sync across threads.
