+++ system/include/emscripten/threading.h
@@ -289,6 +289,8 @@ void emscripten_check_blocking_allowed(void);
 // Experimental API for syncing loaded code between pthreads.
 void _emscripten_thread_sync_code();
 
+void _emscripten_yield();
+
 #ifdef __cplusplus
 }
 #endif
