+++ tests/other/metadce/minimal_main_Oz_USE_PTHREADS_PROXY_TO_PTHREAD.size
@@ -1,1 +1,1 @@
-18576
+18582
